Explicit casts for status
proto_lbs_client.c:
proto_kvlds_client.c:
    be32dec() gives us a uint32_t, and we already checked that it's not
    larger than 1.  So it must be 0 or 1, which can safely fit into an int.

proto_dynamodb_kv_client.c:
    status is exactly 1 or 2, so it's safe to cast to an (int).

+++ lib/proto_kvlds/proto_kvlds_client.c
@@ -155,7 +155,7 @@ callback_donep(void * cookie, uint8_t * buf, size_t buflen)
 			BAD(C->type, "bogus length");
 		if (be32dec(&buf[0]) > 1)
 			BAD(C->type, "bogus status code");
-		status = be32dec(&buf[0]);
+		status = (int)be32dec(&buf[0]);
 
 		/* We successfully parsed this response. */
 		failed = 0;
@@ -190,7 +190,7 @@ callback_get(void * cookie, uint8_t * buf, size_t buflen)
 			BAD("GET", "bogus length");
 		if (be32dec(&buf[0]) > 1)
 			BAD("GET", "bogus status code");
-		status = be32dec(&buf[0]);
+		status = (int)be32dec(&buf[0]);
 
 		/* Parse a value if one exists. */
 		if (status == 0) {
